pyproject.toml
[tool.poetry]
name = "eduboook-python"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.112.2"
uvicorn = "^0.30.6"
firebase-admin = "^6.5.0"
python-jose = "^3.3.0"
python-dotenv = "^1.0.1"
pillow = "^10.4.0"
diffusers = "^0.30.1"
transformers = "^4.44.2"
openai = "^1.43.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

